fix: remove duplicate to separate component
