+++ examples/ui/frontend/src/app/creations/page.tsx
@@ -1,22 +1,20 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { getArtifacts, deleteArtifact, updateArtifact, ArtifactResponse, ArtifactsListResponse } from "@/lib/api/artifacts";
 import { format } from "date-fns";
-import { ArrowLeft, Trash2, Edit, Globe, Copy, Share2 } from "lucide-react";
+import { Trash2, Edit, Globe, Share2 } from "lucide-react";
 import { toast } from "react-hot-toast";
 import ArtifactViewer from "@/components/artifact-viewer";
 import DeleteConfirmationDialog from "@/components/delete-confirmation-dialog";
 import EditNameDialog from "@/components/edit-name-dialog";
-import UserMenu from "@/components/user-menu";
+import CreationsHeader from "@/components/creations-header";
 import UpgradeModal from "@/components/upgrade-modal";
 
 export default function CreationsPage() {
-  const router = useRouter();
   const [artifacts, setArtifacts] = useState<ArtifactResponse[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
@@ -207,39 +205,7 @@ export default function CreationsPage() {
   if (loading) {
     return (
       <div className="min-h-screen bg-gray-50">
-        {/* Header - positioned absolutely over content */}
-        <div className="glass-header p-4 fixed top-0 left-0 right-0 z-10 backdrop-blur-xl bg-white/80">
-          <div className="max-w-4xl mx-auto flex items-center justify-between">
-            <div className="flex items-center space-x-3">
-              <div className="w-10 h-10 bg-gradient-to-r from-blue-500 to-indigo-600 rounded-full flex items-center justify-center">
-                <span className="text-2xl select-none">🐼</span>
-              </div>
-              <div>
-                <h1 className="text-xl font-semibold text-gray-900">
-                  My Creations
-                </h1>
-                <p className="text-sm text-gray-500">
-                  View and manage your saved creations
-                </p>
-              </div>
-            </div>
-
-            <div className="flex items-center space-x-3">
-              {/* Back to Chat Button */}
-              <Button
-                variant="outline"
-                onClick={() => router.push('/chat')}
-                className="flex items-center space-x-2 px-4 py-2 text-sm bg-white hover:bg-gray-50 border border-gray-200 rounded-lg transition-all duration-200 shadow-sm hover:shadow-md"
-              >
-                <ArrowLeft className="w-4 h-4" />
-                <span>Back to Chat</span>
-              </Button>
-
-              {/* User Menu Dropdown */}
-              <UserMenu onUpgradeClick={() => setShowUpgradeModal(true)} />
-            </div>
-          </div>
-        </div>
+        <CreationsHeader onUpgradeClick={() => setShowUpgradeModal(true)} />
 
         {/* Main content with top padding for fixed header */}
         <div className="pt-20 px-6">
@@ -300,39 +266,7 @@ export default function CreationsPage() {
   if (error) {
     return (
       <div className="min-h-screen bg-gray-50">
-        {/* Header - positioned absolutely over content */}
-        <div className="glass-header p-4 fixed top-0 left-0 right-0 z-10 backdrop-blur-xl bg-white/80">
-          <div className="max-w-4xl mx-auto flex items-center justify-between">
-            <div className="flex items-center space-x-3">
-              <div className="w-10 h-10 bg-gradient-to-r from-blue-500 to-indigo-600 rounded-full flex items-center justify-center">
-                <span className="text-2xl select-none">🐼</span>
-              </div>
-              <div>
-                <h1 className="text-xl font-semibold text-gray-900">
-                  My Creations
-                </h1>
-                <p className="text-sm text-gray-500">
-                  View and manage your saved creations
-                </p>
-              </div>
-            </div>
-
-            <div className="flex items-center space-x-3">
-              {/* Back to Chat Button */}
-              <Button
-                variant="outline"
-                onClick={() => router.push('/chat')}
-                className="flex items-center space-x-2 px-4 py-2 text-sm bg-white hover:bg-gray-50 border border-gray-200 rounded-lg transition-all duration-200 shadow-sm hover:shadow-md"
-              >
-                <ArrowLeft className="w-4 h-4" />
-                <span>Back to Chat</span>
-              </Button>
-
-              {/* User Menu Dropdown */}
-              <UserMenu onUpgradeClick={() => setShowUpgradeModal(true)} />
-            </div>
-          </div>
-        </div>
+        <CreationsHeader onUpgradeClick={() => setShowUpgradeModal(true)} />
 
         {/* Main content with top padding for fixed header */}
         <div className="pt-20 px-6">
@@ -366,39 +300,7 @@ export default function CreationsPage() {
 
   return (
     <div className="min-h-screen bg-gray-50">
-      {/* Header - positioned absolutely over content */}
-      <div className="glass-header p-4 fixed top-0 left-0 right-0 z-10 backdrop-blur-xl bg-white/80">
-        <div className="max-w-4xl mx-auto flex items-center justify-between">
-          <div className="flex items-center space-x-3">
-            <div className="w-10 h-10 bg-gradient-to-r from-blue-500 to-indigo-600 rounded-full flex items-center justify-center">
-              <span className="text-2xl select-none">🐼</span>
-            </div>
-            <div>
-              <h1 className="text-xl font-semibold text-gray-900">
-                My Creations
-              </h1>
-              <p className="text-sm text-gray-500">
-                View and manage your saved creations
-              </p>
-            </div>
-          </div>
-
-          <div className="flex items-center space-x-3">
-            {/* Back to Chat Button */}
-            <Button
-              variant="outline"
-              onClick={() => router.push('/chat')}
-              className="flex items-center space-x-2 px-4 py-2 text-sm bg-white hover:bg-gray-50 border border-gray-200 rounded-lg transition-all duration-200 shadow-sm hover:shadow-md"
-            >
-              <ArrowLeft className="w-4 h-4" />
-              <span>Back to Chat</span>
-            </Button>
-
-            {/* User Menu Dropdown */}
-            <UserMenu onUpgradeClick={() => setShowUpgradeModal(true)} />
-          </div>
-        </div>
-      </div>
+      <CreationsHeader onUpgradeClick={() => setShowUpgradeModal(true)} />
 
       {/* Main content with top padding for fixed header */}
       <div className="pt-20 px-6">
